pyproject.toml
[tool.poetry]
name = "truckgen-ai-backend"
version = "0.1.0"
description = "Image generation backend for TruckGen AI application"
authors = ["JineshPrajapat <prajapatjinesh585@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.124.4"
uvicorn = "^0.38.0"
sqlalchemy = "^2.0.45"
psycopg2-binary = "^2.9.11"
alembic = "^1.17.2"
pydantic = {extras = ["email"], version = "^2.12.5"}
python-dotenv = "^1.2.1"
requests = "^2.32.5"
openai = "^2.13.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "==3.2.2"
python-jose = "^3.5.0"
orjson = "^3.11.5"
redis = "^7.1.0"
pydantic-settings = "^2.12.0"
pgvector = "^0.4.2"
psycopg = "^3.3.2"
langchain = "^1.2.0"
sentence-transformers = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
